+++ src/Facility.CodeGen.CSharp/CSharpGeneratorSettings.cs
@@ -18,6 +18,11 @@ public sealed class CSharpGeneratorSettings : FileGeneratorSettings
 	public bool UseNullableReferences { get; set; }
 
 	/// <summary>
+	/// True if C# names should automatically use PascalCase instead of snake case.
+	/// </summary>
+	public bool FixSnakeCase { get; set; }
+
+	/// <summary>
 	/// True to support MessagePack serialization.
 	/// </summary>
 	public bool SupportMessagePack { get; set; }