+++ src/Facility.Core/Http/HttpClientService.cs
@@ -178,7 +178,7 @@ public abstract class HttpClientService
 	/// </summary>
 	protected virtual bool ShouldCreateErrorFromException(Exception exception)
 	{
-		if (exception is ArgumentException || exception is ObjectDisposedException || exception is AggregateException)
+		if (exception is ArgumentException or ObjectDisposedException or AggregateException or UriFormatException)
 			return true;
 
 		var exceptionTypeName = exception.GetType().FullName;
@@ -242,7 +242,7 @@ public abstract class HttpClientService
 #if NET6_0_OR_GREATER
 					url = string.Concat(url.AsSpan(0, bracketedKeyIndex), Uri.EscapeDataString(parameter.Value), url.AsSpan(bracketedKeyIndex + bracketedKey.Length));
 #else
-						url = url.Substring(0, bracketedKeyIndex) + Uri.EscapeDataString(parameter.Value) + url.Substring(bracketedKeyIndex + bracketedKey.Length);
+					url = url.Substring(0, bracketedKeyIndex) + Uri.EscapeDataString(parameter.Value) + url.Substring(bracketedKeyIndex + bracketedKey.Length);
 #endif
 				}
 				else